fix build system, modularize app, add gitignore
It builds once again!

The changes were lumped all together, so here is a log:

Fixed the build system by making the app portion modularized.
Initially the root `build.gradle` was being improperly used for
defining app dependencies. (The project would not build for me.)

`cleanser` was renamed to `Cleanser`

Changed the package name from come.example.myapplication to
com.keturahedwards.cleanskin, this can be changed to something else
in the future if needed. For example, com.keturahdev.cleanskin and
friends.

Added a .gitignore for Android Studio.

On device testing was done and I was unable to scroll the view,
this might be a regression so do not merge this into master. If this
version builds and functions properly on your device this likely means
there is some legacy issue support.

Tested on Android version: 6.0.1